ADDED: To a File

+++ src/main/java/EmployeePayroll.java
@@ -1,9 +1,9 @@
 public class EmployeePayroll {
     int id;
     String name;
-    int salary;
+    double salary;
 
-    public EmployeePayroll(int id, String name, int salary) {
+    public EmployeePayroll(int id, String name, double salary) {
         this.id = id;
         this.name = name;
         this.salary = salary;
